Add deploy key functionality
