+++ Cargo.toml
@@ -5,6 +5,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+clap = { version = "4.5.28", features = ["derive"] }
 colored = "2.0.0"
 csv = "1.3"
 reqwest = { version = "0.11.6", features = ["blocking", "json"] }
